refactor: Home 페이지를 12기 모집 페이지로 바꾸기 위해 Home 관련 컴포넌트 컨테이너로 감싸기
