+++ src/app/page.tsx
@@ -1,17 +1,9 @@
-import CurrentlySection from "@/app/_components/CurrentlySection";
-import AboutUsSection from "@/app/_components/AboutUsSection";
-import WhatWeDoSection from "@/app/_components/WhatWeDoSection";
-import ProjectSection from "@/app/_components/ProjectSection";
-import CrewMembersSection from "@/app/_components/CrewMembersSection";
+import HomeContainer from "@/app/_components/Home";
 
 export default function Home() {
   return (
     <>
-      <CurrentlySection />
-      <AboutUsSection />
-      <WhatWeDoSection />
-      <ProjectSection />
-      <CrewMembersSection />
+      <HomeContainer />
     </>
   );
 }
